% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bus.R
\name{tm_bus_routes}
\alias{tm_bus_routes}
\title{List all available bus routes}
\usage{
tm_bus_routes()
}
\value{
A sorted character vector of bus route IDs.
}
\description{
List all available bus routes
}
\examples{
tm_bus_routes()
}
