% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_bus_stations}
\alias{tm_bus_stations}
\title{List all stations on a bus route}
\usage{
tm_bus_stations(route)
}
\arguments{
\item{route}{Bus route ID. Use \code{\link[=tm_bus_routes]{tm_bus_routes()}} to see all valid IDs.}
}
\value{
A data frame with columns \code{stop_name}, \code{station}, and \code{order}.
}
\description{
List all stations on a bus route
}
\examples{
tm_bus_stations("1")
}
