% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_bus_stop_id}
\alias{tm_bus_stop_id}
\title{Look up stop IDs for a bus route station}
\usage{
tm_bus_stop_id(route, stop_name, direction)
}
\arguments{
\item{route}{Bus route ID, e.g. \code{"1"}, \code{"66"}, \code{"CT2"}, \code{"SL1"}.
Use \code{\link[=tm_bus_routes]{tm_bus_routes()}} to see all valid route IDs.}

\item{stop_name}{Station name (case-insensitive). Use \code{\link[=tm_bus_stations]{tm_bus_stations()}}
to see valid names for a route.}

\item{direction}{\code{"outbound"}, \code{"inbound"}, \code{0}, or \code{1}.}
}
\value{
A character vector of stop IDs.
}
\description{
Look up stop IDs for a bus route station
}
\examples{
tm_bus_stop_id("1", "Harvard", "inbound")
tm_bus_stop_id("66", "Harvard", "outbound")
}
