% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_cr_stations}
\alias{tm_cr_stations}
\title{List all stations on a commuter rail line}
\usage{
tm_cr_stations(route)
}
\arguments{
\item{route}{CR route ID. Use \code{\link[=tm_cr_routes]{tm_cr_routes()}} to see all valid IDs.}
}
\value{
A data frame with columns \code{stop_name}, \code{station} (GTFS place ID),
\code{order}, and \code{terminus}.
}
\description{
List all stations on a commuter rail line
}
\examples{
tm_cr_stations("CR-Fairmount")
}
