% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{tm_facilities}
\alias{tm_facilities}
\title{List MBTA facilities}
\usage{
tm_facilities(base_url = tm_base_url())
}
\arguments{
\item{base_url}{Base URL of the TransitMatters API. Defaults to
\code{getOption("tm_dashboard_base_url")} or the production host.}
}
\value{
A list of facility objects.
}
\description{
List MBTA facilities
}
\examples{
\dontrun{
tm_facilities()
}
}
