% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_ferry_routes}
\alias{tm_ferry_routes}
\title{List all available ferry routes}
\usage{
tm_ferry_routes()
}
\value{
A sorted character vector of ferry route IDs available for stop
lookup.
}
\description{
List all available ferry routes
}
\examples{
tm_ferry_routes()
}
