% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_ferry_stations}
\alias{tm_ferry_stations}
\title{List all stations on a ferry route}
\usage{
tm_ferry_stations(route)
}
\arguments{
\item{route}{Ferry route ID. Use \code{\link[=tm_ferry_routes]{tm_ferry_routes()}} to see all valid IDs.}
}
\value{
A data frame with columns \code{stop_name}, \code{station}, \code{order},
and \code{terminus}.
}
\description{
List all stations on a ferry route
}
\examples{
tm_ferry_stations("Boat-F1")
}
