% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{tm_git_id}
\alias{tm_git_id}
\title{Get the deployed git commit ID}
\usage{
tm_git_id(base_url = tm_base_url())
}
\arguments{
\item{base_url}{Base URL of the TransitMatters API. Defaults to
\code{getOption("tm_dashboard_base_url")} or the production host.}
}
\value{
A list with the git commit SHA of the running server.
}
\description{
Get the deployed git commit ID
}
\examples{
\dontrun{
tm_git_id()
}
}
