% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{tm_service_ridership_dashboard}
\alias{tm_service_ridership_dashboard}
\title{Get the service ridership dashboard summary}
\usage{
tm_service_ridership_dashboard(base_url = tm_base_url())
}
\arguments{
\item{base_url}{Base URL of the TransitMatters API. Defaults to
\code{getOption("tm_dashboard_base_url")} or the production host.}
}
\value{
A list containing the service ridership dashboard data.
}
\description{
Get the service ridership dashboard summary
}
\examples{
\dontrun{
tm_service_ridership_dashboard()
}
}
