% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{package}
\name{transitmattr-package}
\alias{transitmattr}
\alias{transitmattr-package}
\title{transitmattr: R Client for the TransitMatters Data Dashboard API}
\description{
A thin 'httr2'-based wrapper around the TransitMatters Data Dashboard API, providing access to MBTA performance metrics including headways, dwell times, travel times, ridership, and service data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://transitmatters.github.io/transitmattr/}
  \item \url{https://github.com/transitmatters/transitmattr}
  \item Report bugs at \url{https://github.com/transitmatters/transitmattr/issues}
}

}
\author{
\strong{Maintainer}: Andrew Kouré \email{akoure@transitmatters.info}

Authors:
\itemize{
  \item Andrew Kouré \email{akoure@transitmatters.info}
}

}
\keyword{internal}
