% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{tm_bus_routes}
\alias{tm_bus_routes}
\title{List all available bus routes}
\usage{
tm_bus_routes()
}
\value{
A sorted character vector of bus route IDs available for stop lookup.
}
\description{
List all available bus routes
}
\examples{
tm_bus_routes()
}
