% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cr.R
\name{tm_cr_routes}
\alias{tm_cr_routes}
\title{List all available commuter rail routes}
\usage{
tm_cr_routes()
}
\value{
A sorted character vector of CR route IDs.
}
\description{
List all available commuter rail routes
}
\examples{
tm_cr_routes()
}
