% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{tm_time_predictions}
\alias{tm_time_predictions}
\title{Get current time predictions}
\usage{
tm_time_predictions(base_url = tm_base_url())
}
\arguments{
\item{base_url}{Base URL of the TransitMatters API. Defaults to
\code{getOption("tm_dashboard_base_url")} or the production host.}
}
\value{
A list of time prediction objects.
}
\description{
Get current time predictions
}
\examples{
\dontrun{
tm_time_predictions()
}
}
